add command interactions with db

+++ server/src/main/java/org/example/db/DatabaseRequests.java
@@ -4,36 +4,36 @@ public class DatabaseRequests {
     public static String createAllTables() {
         return """
                 DO $$ BEGIN
-                CREATE TYPE color AS ENUM(
+                    CREATE TYPE color AS ENUM(
                         'GREEN',
                         'BLACK',
                         'WHITE',
                         'RED',
                         'BLUE',
                         'ORANGE'
-                );
-                CREATE TYPE country AS ENUM(
+                    );
+                    CREATE TYPE country AS ENUM(
                         'RUSSIA',
                         'USA',
                         'ITALY',
                         'THAILAND',
-                        'SOUTH_KOREA',
-                        );
-                CREATE TYPE difficulty AS ENUM(
+                        'SOUTH_KOREA'
+                    );
+                    CREATE TYPE difficulty AS ENUM(
                         'VERY_HARD',
                         'IMPOSSIBLE',
                         'HOPELESS'
-                );
+                    );
                 EXCEPTION
-                WHEN duplicate_object THEN null;
+                    WHEN duplicate_object THEN null;
                 END $$;
                 CREATE TABLE IF NOT EXISTS authors (
                         id SERIAL PRIMARY KEY,
                         name TEXT NOT NULL,
                         weight FLOAT NOT NULL,
-                        eye_color COLOR NOT NULL,
-                        hair_color COLOR NOT NULL,
-                        nationality COUNTRY NOT NULL
+                        eye_color color NOT NULL,
+                        hair_color color NOT NULL,
+                        nationality country NOT NULL
                     );
                 CREATE TABLE IF NOT EXISTS users (
                         id SERIAL PRIMARY KEY,
@@ -48,25 +48,40 @@ public class DatabaseRequests {
                         coord_y FLOAT NOT NULL ,
                         creation_date DATE NOT NULL ,
                         minimal_point INTEGER,
-                        difficulty DIFFICULTY NOT NULL,
-                        person_id REFERENCES people(id),
-                        user_id REFERENCES users(id)
+                        difficulty difficulty NOT NULL,
+                        author_id INTEGER REFERENCES authors(id),
+                        user_id INTEGER REFERENCES users(id)
                     );
                 """;
     }
 
     public static String addUser() {
         return """
-                INSERT INTO users(login, password, salt)
-                VALUES (?, ?, ?);
+                INSERT INTO users(username, password, salt)
+                VALUES (?, ?, ?)
                 RETURNING id;
+                """;
+    }
+
+    public static String checkIfAuthorExists() {
+        return """
+                SELECT id FROM authors
+                   WHERE name = ? AND weight = ?
+                        AND eye_color = ? AND hair_color = ?
+                        AND nationality = ?;
+                """;
+    }
+
+    public static String countLabWorks() {
+        return """
+                SELECT COUNT(*) FROM lab_works WHERE (user_id = ?);
                 """;
     }
 
     public static String addAuthor() {
         return """
-                INSERT INTO people(name, weight, height, eye_color, heir_color, nationality)
-                VALUES (?, ?, ?, ?, ?, ?)
+                INSERT INTO authors(name, weight, eye_color, hair_color, nationality)
+                VALUES (?, ?, ?, ?, ?)
                 RETURNING id;
                 """;
     }
@@ -85,6 +100,12 @@ public class DatabaseRequests {
                 """;
     }
 
+    public static String getLabWOrkById() {
+        return """
+                SELECT * FROM lab_works WHERE (id = ?);
+                """;
+    }
+
     public static String getUserByUsername() {
         return """
                 SELECT * FROM users WHERE (username = ?);
@@ -99,19 +120,27 @@ public class DatabaseRequests {
 
     public static String getAllLabWorks() {
         return """
-                SELECT * FROM lab_works;
+                SELECT * FROM lab_works WHERE (user_id = ?);
                 """;
     }
 
-    public static String getAllLabWorksJoinByAuthorId() {
+    public static String getAllLabWorksJoinAuthorIdByUserId() {
         return """
-                SELECT * FROM lab_works JOIN authors ON lab_works.author_id = authors.id;
+                SELECT * FROM lab_works
+                JOIN authors ON lab_works.author_id = authors.id
+                WHERE (lab_works.user_id = ?);
                 """;
     }
 
-    public static String deleteLabWorkByUserId() {
+    public static String deleteAllLabWorkByUserId() {
         return """
-                DELETE FROM lab_works WHERE (user_id = ?);
+                DELETE FROM lab_works WHERE (user_id = ?) RETURNING user_id;
+                """;
+    }
+
+    public static String deleteLabWorkById() {
+        return """
+                DELETE FROM lab_works WHERE id = ? AND user_id = ? RETURNING id;
                 """;
     }
 
@@ -120,7 +149,7 @@ public class DatabaseRequests {
                 UPDATE lab_works
                 SET (name, coord_x, coord_y, creation_date, minimal_point, difficulty)
                  = (?, ?, ?, ?, ?, ?)
-                WHERE (id = ?) AND (user_id = ?)
+                WHERE id = ? AND user_id = ?
                 RETURNING id;
                 """;
     }
